fix import bug

+++ payrollapp/models.py
@@ -7,10 +7,10 @@ class Payroll(models.Model):
         pass
 
 class Paycheck(models.Model):
-    amount = models.FloatField(max_digits=8, decimal_places=2, default=0)
+    amount = models.DecimalField(max_digits=8, decimal_places=2, default=0)
 
 class PurchaseOrder(models.Model):
-    amount = models.FloatField(max_digits=10, decimal_places=2, default=0)
+    amount = models.DecimalField(max_digits=10, decimal_places=2, default=0)
